refactor(wildcat-babel): Consolidate Babel transpilation with babelDevTranspiler
Keeping things DRY and expectant by combining this functionality with babelDevTranspiler
