+++ packages/react-wildcat/src/middleware/babelDevTranspiler.js
@@ -1,9 +1,10 @@
 const fs = require("fs-extra");
 const path = require("path");
-const chalk = require("chalk");
-const minimatch = require("minimatch");
 const pathExists = require("path-exists");
 const pathResolve = require("resolve-path");
+
+const prettyLog = require("../utils/prettyLog");
+const transpile = require("../utils/transpile");
 
 module.exports = function babelDevTranspiler(root, options) {
     "use strict";
@@ -21,156 +22,37 @@ module.exports = function babelDevTranspiler(root, options) {
     const outDir = options.outDir;
     const sourceDir = options.sourceDir;
 
-    let instrumenter, instrumentationExcludes;
-
-    if (coverage) {
-        const coverageEnv = coverageSettings.env;
-        const coverageEnvSettings = coverageSettings[coverageEnv] || {};
-        const instrumentationSettings = coverageEnvSettings.instrumentation || {};
-
-        instrumentationExcludes = instrumentationSettings.excludes || [];
-
-        const coverageFiles = process.env.COVERAGE_FILES;
-        const suite = process.env.COVERAGE_SUITE;
-
-        if (coverageFiles) {
-            // Istanbul only supports excluding files,
-            // so the value of files must be converted to a negated exclusion
-            // i.e. --cover-files=foo.js,bar.js -> !{foo.js,bar.js}
-            const excludedFiles = [
-                coverageFiles.includes(",") ? `!{${coverageFiles.trim()}}` : `!${coverageFiles.trim()}`
-            ];
-
-            instrumentationExcludes = excludedFiles;
-        } else if (suite && instrumentationSettings.onSuiteExcludeCoverage) {
-            instrumentationExcludes = instrumentationExcludes.concat(
-                instrumentationSettings.onSuiteExcludeCoverage(suite)
-            );
-        }
-
-        const Istanbul = require("istanbul");
-        instrumenter = new Istanbul.Instrumenter(instrumentationSettings);
-    }
-
-    function addSourceMappingUrl(code, loc) {
-        return code + "\n//# sourceMappingURL=" + path.basename(loc);
-    }
-
     function* _babelDevTranspiler(ctx, pathOptions) {
-        const babel = require("babel-core");
+        "use strict";
 
         const modulePath = pathOptions.modulePath;
         const moduleSourcePath = pathOptions.moduleSourcePath;
         const moduleBinPath = pathOptions.moduleBinPath;
-        const relativePath = pathOptions.relativePath;
+
+        if (temporaryCache.has(modulePath)) {
+            return temporaryCache.get(modulePath);
+        }
 
         const dataOptions = Object.assign({}, babelOptions, {
             sourceFileName: moduleSourcePath,
             sourceMapTarget: path.basename(moduleSourcePath)
         });
 
-        const statusCode = chalk.styles.yellow.open + 201 + chalk.styles.yellow.close;
-        const relOut = chalk.styles.grey.open + modulePath.replace(`${root}`, "") + chalk.styles.grey.close;
-        const prettyLog = `${statusCode} CREATE ${relOut}`;
-
-        if (temporaryCache.has(modulePath)) {
-            return temporaryCache.get(modulePath);
-        }
-
         const transpilerPromise = new Promise(function transpilerPromise(resolve, reject) {
             if (extensions.indexOf(path.extname(moduleSourcePath)) !== -1) {
-                babel.transformFile(moduleSourcePath, dataOptions, function transformFile(transformErr, data) {
-                    if (transformErr) {
-                        logger.error(transformErr);
-                        return reject(transformErr);
-                    }
+                transpile({
+                    coverage,
+                    coverageSettings,
 
-                    if (data.ignored) {
-                        return resolve();
-                    }
+                    babelOptions,
+                    dataOptions,
+                    pathOptions,
 
-                    const sourceMaps = babelOptions.sourceMaps || ((babelOptions.env || {}).development || {}).sourceMaps;
+                    logger,
+                    logLevel,
 
-                    if (data.map && sourceMaps && sourceMaps !== "inline") {
-                        const mapLoc = modulePath + ".map";
-                        data.code = addSourceMappingUrl(data.code, mapLoc);
-
-                        fs.createOutputStream(mapLoc)
-                            .on("open", function outputStreamOpen() {
-                                if (logLevel > 1) {
-                                    logger.meta(prettyLog);
-                                }
-                            })
-                            .on("error", function outputStreamError(outputErr) {
-                                logger.error(outputErr);
-                            })
-                            .end(JSON.stringify(data.map));
-                    }
-
-                    let instrumentedCodePromise;
-
-                    if (
-                        coverage &&
-                        instrumenter &&
-                        !instrumentationExcludes.some(minimatch.bind(null, relativePath))
-                    ) {
-                        instrumentedCodePromise = new Promise(
-                            function instrumenterPromise(resolveInstrumenter, rejectInstrumenter) {
-                                instrumenter.instrument(
-                                    data.code,
-                                    relativePath,
-                                    function instrumentOutput(instrumentError, instrumentedCode) {
-                                        if (instrumentError) {
-                                            return rejectInstrumenter(instrumentError);
-                                        }
-
-                                        return resolveInstrumenter(instrumentedCode);
-                                    }
-                                );
-                            }
-                        );
-                    } else {
-                        instrumentedCodePromise = Promise.resolve(data.code);
-                    }
-
-                    instrumentedCodePromise
-                        .then(function instrumentedCodeResult(code) {
-                            const res = {
-                                type: "application/x-es-module",
-                                length: code.length,
-                                body: code,
-                                status: 200
-                            };
-
-                            fs.createOutputStream(modulePath)
-                                .on("open", function outputStreamOpen() {
-                                    if (logLevel > 1) {
-                                        logger.meta(prettyLog);
-                                    }
-                                })
-                                .on("error", function outputStreamError(outputErr) {
-                                    // Disk write error, delete the temporary cache
-                                    temporaryCache.delete(modulePath);
-                                    logger.error(outputErr);
-                                })
-                                .on("finish", function outputStreamFinish() {
-                                    // Disk write complete, delete the temporary cache
-                                    temporaryCache.delete(modulePath);
-                                })
-                                .end(code);
-
-                            // Code is available before disk write, update the temporary cache
-                            temporaryCache.set(modulePath, Promise.resolve(res));
-                            return resolve(res);
-                        })
-                        .catch(function instrumentedCodeError(err) {
-                            // Disk write error, delete the temporary cache
-                            temporaryCache.delete(modulePath);
-
-                            logger.error(err);
-                            return reject(err);
-                        });
-                });
+                    temporaryCache
+                }, resolve, reject);
             } else {
                 const importable = `module.exports = "${origin}${moduleBinPath.replace(`${root}`, "")}";`;
 
@@ -200,7 +82,7 @@ module.exports = function babelDevTranspiler(root, options) {
                                 fs.createOutputStream(moduleBinPath || modulePath)
                                     .on("open", function binaryStreamOpen() {
                                         if (logLevel > 1) {
-                                            logger.meta(prettyLog);
+                                            logger.meta(prettyLog(modulePath));
                                         }
                                     })
                                     .on("error", function binaryStreamError(outputErr) {
@@ -231,7 +113,9 @@ module.exports = function babelDevTranspiler(root, options) {
         return transpilerPromise;
     }
 
-    return function* transpile(next) {
+    return function* setupTranspiler(next) {
+        "use strict";
+
         const request = this.request;
         const res = this.response;
