+++ packages/react-wildcat/src/utils/prettyLog.js
@@ -1,0 +1,8 @@
+const chalk = require("chalk");
+
+module.exports = function prettyLog(modulePath) {
+    const statusCode = chalk.styles.yellow.open + 201 + chalk.styles.yellow.close;
+    const relOut = chalk.styles.grey.open + modulePath.replace(`${root}`, "") + chalk.styles.grey.close;
+
+    return `${statusCode} CREATE ${relOut}`;
+};
